requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
django-redis==4.12.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Pillow==8.1.0
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2021.1
redis==3.5.3
sqlparse==0.4.1
